+++ test/postgresql/sql/select_builder_test.dart
@@ -4,7 +4,7 @@ import 'package:test/test.dart';
 import 'package:cl_datahub/src/persistence/postgresql/sql/sql.dart';
 
 void main() {
-  final fieldX = DataField(FieldType.String, 'fieldX');
+  final fieldX = DataField(FieldType.String, 'fake', 'fieldX');
   final schemaTable = SelectFromTable('schema', 'table');
   final otherTable = SelectFromTable('schema', 'other');
   final otherTable2 = SelectFromTable('schema', 'different');
@@ -21,7 +21,7 @@ void main() {
       'Select filter eq string',
       _test(
         SelectBuilder(schemaTable)..where(Filter.equals(fieldX, 'valueX')),
-        'SELECT * FROM schema.table WHERE "fieldX" = \'valueX\'',
+        'SELECT * FROM schema.table WHERE "fake"."fieldX" = \'valueX\'',
       ),
     );
 
@@ -31,7 +31,7 @@ void main() {
         SelectBuilder(schemaTable)
           ..where(PropertyCompare(PropertyCompareType.Equals, fieldX, 'valueX',
               caseSensitive: false)),
-        'SELECT * FROM schema.table WHERE LOWER("fieldX") = \'valuex\'',
+        'SELECT * FROM schema.table WHERE LOWER("fake"."fieldX") = \'valuex\'',
       ),
     );
 
@@ -41,7 +41,7 @@ void main() {
         SelectBuilder(schemaTable)
           ..where(
               PropertyCompare(PropertyCompareType.Contains, fieldX, 'valueX')),
-        'SELECT * FROM schema.table WHERE "fieldX" LIKE \'%valueX%\'',
+        'SELECT * FROM schema.table WHERE "fake"."fieldX" LIKE \'%valueX%\'',
       ),
     );
 
@@ -52,7 +52,7 @@ void main() {
           ..where(PropertyCompare(
               PropertyCompareType.Contains, fieldX, 'valueX',
               caseSensitive: false)),
-        'SELECT * FROM schema.table WHERE "fieldX" ILIKE \'%valueX%\'',
+        'SELECT * FROM schema.table WHERE "fake"."fieldX" ILIKE \'%valueX%\'',
       ),
     );
 
@@ -60,7 +60,7 @@ void main() {
       'Select filter eq int',
       _test(
         SelectBuilder(schemaTable)..where(Filter.equals(fieldX, 20)),
-        'SELECT * FROM schema.table WHERE "fieldX" = 20',
+        'SELECT * FROM schema.table WHERE "fake"."fieldX" = 20',
       ),
     );
 
@@ -68,7 +68,7 @@ void main() {
       'Select filter eq double',
       _test(
         SelectBuilder(schemaTable)..where(Filter.equals(fieldX, 20.12)),
-        'SELECT * FROM schema.table WHERE "fieldX" = 20.12',
+        'SELECT * FROM schema.table WHERE "fake"."fieldX" = 20.12',
       ),
     );
   });
@@ -78,14 +78,14 @@ void main() {
       'Join other',
       _test(
         SelectBuilder(
-          JoinedSelectSource(
+          JoinedSelectFrom(
             schemaTable,
             [
               TableJoin(otherTable, 'id', PropertyCompareType.Equals, 'main_id')
             ],
           ),
         ),
-        'SELECT * FROM schema.table JOIN schema.other ON schema.table.id = schema.other.main_id',
+        'SELECT * FROM schema.table JOIN schema.other ON schema.table."id" = schema.other."main_id"',
       ),
     );
 
@@ -93,15 +93,15 @@ void main() {
       'Join other filter eq string',
       _test(
         SelectBuilder(
-          JoinedSelectSource(
+          JoinedSelectFrom(
             schemaTable,
             [
               TableJoin(otherTable, 'id', PropertyCompareType.Equals, 'main_id')
             ],
           ),
         )..where(Filter.equals(fieldX, 'valueX')),
-        'SELECT * FROM schema.table JOIN schema.other ON schema.table.id = '
-        'schema.other.main_id WHERE "fieldX" = \'valueX\'',
+        'SELECT * FROM schema.table JOIN schema.other ON schema.table."id" = '
+        'schema.other."main_id" WHERE "fake"."fieldX" = \'valueX\'',
       ),
     );
 
@@ -109,7 +109,7 @@ void main() {
       'Join multiple',
       _test(
         SelectBuilder(
-          JoinedSelectSource(
+          JoinedSelectFrom(
             schemaTable,
             [
               TableJoin(
@@ -120,9 +120,9 @@ void main() {
           ),
         )..where(PropertyCompare(PropertyCompareType.Equals, fieldX, 'valueX',
             caseSensitive: false)),
-        'SELECT * FROM schema.table JOIN schema.other ON schema.table.id = '
-        'schema.other.main_id JOIN schema.different ON schema.table.xyz < '
-        'schema.different.abc WHERE LOWER("fieldX") = \'valuex\'',
+        'SELECT * FROM schema.table JOIN schema.other ON schema.table."id" = '
+        'schema.other."main_id" JOIN schema.different ON schema.table."xyz" < '
+        'schema.different."abc" WHERE LOWER("fake"."fieldX") = \'valuex\'',
       ),
     );
   });